% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoga_code.R
\name{pseudoga}
\alias{pseudoga}
\title{Function to estimate pseudotime}
\usage{
pseudoga(
  sce,
  type = c("counts", "nomralized", "expression"),
  ntest = 50,
  minit = 30,
  epsilon = 1e-04,
  normalization = c("default", "TMM", "quant", "cpm"),
  nnprop = 0.3
)
}
\arguments{
\item{sce}{A SingleCellExpression object. The slot "assays" must contain one of the followings: "count" (count data), "normalized" (normalized data), "expression" (neither count data nor normalized data).}

\item{type}{Type of expression data to be analyzed. It should assume one of these values: "counts", "normalized", "expression". Default is "counts".}

\item{ntest}{Size of each generation considered in genetic algorithm is (8*ntest) . Default value for ntest is 50.}

\item{minit}{Minimum number of generations considered in genetic algorithm. Default value is 30.}

\item{epsilon}{The tolerance value used in convergence. Default value is \eqn{10^{-4}}.}

\item{normalization}{Normalization method applied on the input dataset. It must be one of the following "default" (DESeq Normalization), "TMM" (TMM Normalization), "quant" (Quantile Normalization) or "cpm" (Counts Per Million).}

\item{nnprop}{Proportion of sample size used as value of k for kNN smoothing. If sample size is N, (N*nnprop) is considered as value of k. Default value is 0.3.}
}
\value{
An updated SingleCellExperiment object in which colData contains values for pseudotime for each cell with slot name "Pseudotime" .
}
\description{
This function estimates pseudotime using PseudoGA algorithm. Ordering of cells is performed using genetic algorithm. Subsequently, pseudotimes are updated using nearest neighbor smoothing.
}
\examples{
\dontrun{
library(SingleCellExperiment)
counts <- matrix(rpois(10000, lambda = 10), ncol=100, nrow=100)
sce <- SingleCellExperiment(list(counts=counts))
sce<-pseudoga(sce)
}
}
