% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_genes.R
\name{select_genes}
\alias{select_genes}
\title{Function to perform gene selection}
\usage{
select_genes(
  sce,
  type = c("counts", "nomralized", "expression"),
  numgenes = 500
)
}
\arguments{
\item{sce}{A SingleCellExpression object. The slot "assays" must contain one of the followings: "counts" (count data), "normalized" (normalized data), "expression" (neither count data nor normalized data). The slot "colData" must contain a variable "cluster" with cluster ids of the cells. If no cluster id is provided, the program divides the dataset into two clusters and performs the analysis.}

\item{type}{Type of expression data to be analyzed. It should assume one of these values: "counts", "normalized", "expression". Default is "counts".}

\item{numgenes}{Number of top genes included in the final dataset from each of the differential expression analyses between pairs of clusters. Default value is 500.}
}
\value{
A SingleCellExperiment object with expression matrix containing selected genes.
}
\description{
This function performs gene selection on a single cell expression dataset. Given a dataset containing cell cluster ids, this function performs differential expression analysis between all pairs of clusters. The final gene set is chosen as the union of top differentially expressed genes obtained from all comparisons between pairs of clusters.
}
\examples{
\dontrun{
library(SingleCellExperiment)
counts <- matrix(rpois(3000000, lambda = 10), ncol=300, nrow=10000)
rownames(counts)<-1:(dim(counts)[1])
colnames(counts)<-1:(dim(counts)[2])
sce <- SingleCellExperiment(list(counts=counts))
km<-kmeans(t(counts),3)
colData(sce)$cluster<-km$cluster
sce1<-select_genes(sce)
}
}
